Changes made to the basket page, all products are now pulling through (names etc)
The new checkout page layout, also fixed the product on the pay CTA, the product info pulls through etc.

Added a new 'Products Purchased' Box to the members area (items completed at checkout will be shown here)

Added a new 'Saved for later' box to the members area, you can add the item from the basket with the 'save for later' cta and it will appear in the members area box and vise versa from the members area back to the basket.

Note: I can't make the save for later CTA on the basket look nice, CTA's hate me lol

Note: The ability to 'save for later' is only shown to users that are logged in. The CTA is hidden for non logged in users

Note: When the 'save for later' cta is clicked it will be sent to the members area but also persist on the basket page

Note: When the user clicks the 'save for later' cta they will see a toast msg saying it has been added to their members area

Now when a user is logged in they will be able to access the 'members' page in the nav, I called it 'Members Area'

On the sign in page there is now a CTA/link that allows the user to return to the homepage

On the signup page there is now a CTA/link that allows the user to return to the homepage